refactor: optimize the code
