+++ backend/routes/posts.js
@@ -1,157 +1,28 @@
 const express = require("express");
-const multer = require("multer");
 
-const Post = require('../models/post');
+const extractFile = require("../middleware/file");
 const checkAuth = require("../middleware/check-auth");
+
+const PostController = require("../controllers/post");
 
 const router = express.Router();
 
-const MIME_TYPE_MAP = {
-    'image/png': 'png',
-    'image/jpeg': 'jpg',
-    'image/jpg': 'jpg'
-}
 
-const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    const isValid = MIME_TYPE_MAP[file.mimetype];
-    let error = new Error("Invalid mime type");
-    if(isValid){
-      error = null;
-    }
-    cb(error, "backend/images");
-  },
-  filename: (req, file, cb) => {
-    const name = file.originalname.toLowerCase().split(' ').join('-');
-    const extension = MIME_TYPE_MAP[file.mimetype];
-    cb(null, name + '-' + Date.now() + '-' +extension);
-  }
-});
 
 router.post("",
   checkAuth,
-  multer({storage: storage}).single("image"), (req, res, next) => {
-  const url = req.protocol + '://' + req.get("host");
-  const post = new Post({
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename,
-    creator: req.userData.userId
-  });
-  post.save().then(createdPost => {
-    console.log(createdPost);
-    res.status(201).json({
-      message: 'Post added successfully',
-      post: {
-        ...createdPost,
-        id: createdPost._id
-      }
-    });
-  })
-  .catch(error => {
-    res.status(500).json({
-      message: "Failed to create a post!"
-    });
-  });
-});
+  extractFile, PostController.createPost);
 
 router.put("/:id",
   checkAuth,
-  multer({storage: storage}).single("image"), (req, res, next)=>{
-  let imagePath = req.body.imagePath;
-  if (req.file) {
-    const url = req.protocol + '://' + req.get("host");
-    imagePath = url + "/images/" + req.file.filename;
-  }
-
-  const post = new Post({
-    _id: req.body.id,
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: imagePath,
-    creator: req.userData.userId
-  });
-  console.log(post);
-  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result=>{
-    if (result.nModified > 0) {
-      res.status(200).json({
-        message: 'Post updated successfully'
-      });
-    }
-    res.status(401).json({
-      message: 'You are not authorized to update this post!'
-    });
-  })
-  .catch(error => {
-    res.status(500).json({
-      message: "Failed to update the post!"
-    });
-  });
-});
+  extractFile, PostController.updatePostById);
 
 //get all posts
-router.get("", (req, res, next) => {
-  const pageSize = +req.query.pageSize;
-  const currentPage = req.query.page;
-  const postQuery = Post.find();
-  let fetchedPosts;
-
-  if (pageSize&&currentPage){
-    postQuery.skip(pageSize*(currentPage-1)).limit(pageSize);
-  }
-  //then() chain queries
-  postQuery.then((documents) => {
-    console.log(documents);
-    fetchedPosts = documents;
-    //send back response to client
-    return Post.estimatedDocumentCount();
-  })
-  .then(count => {
-    res.status(200).json({
-      message: 'Post fetched successfully',
-      posts: fetchedPosts,
-      maxPosts: count
-    });
-  })
-  .catch(error => {
-    res.status(500).json({
-      message: "Failed to fetch posts!"
-    });
-  });
-});
+router.get("", PostController.getAllPosts);
 
 //get post by id
-router.get("/:id", (req, res, next) => {
-  Post.findById(req.params.id).then(post => {
-    if (post) {
-      res.status(200).json(post);
-    } else {
-      res.status(404).json({message: "Post not found!"});
-    }
-  }).catch(error => {
-    res.status(500).json({
-      message: "Failed to find the post!"
-    });
-  });
-});
+router.get("/:id", PostController.getPostById);
 
-router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
-    if (result.n > 0) {
-      res.status(200).json({
-        message: 'Post delete successfully!'
-      });
-    }
-    res.status(401).json({
-      message: 'You are not authorized to delete this post!'
-    })
-    .catch(error => {
-      res.status(500).json({
-        message: "Failed to delete the post!"
-      });
-    });
-  });
-
-});
+router.delete("/:id", checkAuth, PostController.deletePostById);
 
 module.exports = router;
