+++ src/app/Post/posts.module.ts
@@ -1,0 +1,24 @@
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { ReactiveFormsModule } from '@angular/forms';
+import { RouterModule } from "@angular/router";
+
+import { AngularMetraialModule } from "../angular-matrial.module";
+
+import { PostCreateComponent } from './post-create/post-create.component';
+import { PostListComponent } from './post-list/post-list.component';
+
+@NgModule({
+  declarations: [
+    PostCreateComponent,
+    PostListComponent,
+  ],
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    AngularMetraialModule,
+    RouterModule
+  ]
+})
+
+export class PostsModule{}
