+++ src/app/angular-matrial.module.ts
@@ -1,0 +1,27 @@
+import { NgModule } from "@angular/core";
+
+import { MatInputModule } from '@angular/material/input';
+import { MatCardModule } from '@angular/material/card';
+import { MatButtonModule } from '@angular/material/button';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatExpansionModule } from '@angular/material/expansion';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatIconModule } from '@angular/material/icon';
+import { MatDialogModule } from '@angular/material/dialog';
+
+@NgModule({
+  exports: [
+    MatInputModule,
+    MatCardModule,
+    MatButtonModule,
+    MatToolbarModule,
+    MatExpansionModule,
+    MatDialogModule,
+    MatProgressSpinnerModule,
+    MatPaginatorModule,
+    MatIconModule
+  ]
+})
+
+export class AngularMetraialModule {}
