+++ src/app/app.module.ts
@@ -1,57 +1,33 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MatInputModule } from '@angular/material/input';
-import { MatCardModule } from '@angular/material/card';
-import { MatButtonModule } from '@angular/material/button';
-import { MatToolbarModule } from '@angular/material/toolbar';
-import { MatExpansionModule } from '@angular/material/expansion';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { MatPaginatorModule } from '@angular/material/paginator';
-import { MatIconModule } from '@angular/material/icon';
 
 import { AppComponent } from './app.component';
-import { PostCreateComponent } from './Post/post-create/post-create.component';
 import { HeaderComponent } from './header/header.component';
-import { PostListComponent } from './Post/post-list/post-list.component';
 import { AppRoutingModule } from './app-routing.module';
 
-import { PostService } from './Post/post-service/post.service';
-import { LoginComponent } from './auth/login/login.component';
-import { SignupComponent } from './auth/signup/signup.component';
 import { AuthInterceptor } from './auth/auth-interceptor';
 import { ErrorInterceptor } from './error-interceptor';
-import { MatDialogModule } from '@angular/material/dialog';
 import { ErrorComponent } from './error/error.component';
+import { AngularMetraialModule } from './angular-matrial.module';
+import { PostsModule } from './Post/posts.module';
+import { AuthModule } from './auth/auth.module';
 
 @NgModule({
   declarations: [
     AppComponent,
-    PostCreateComponent,
     HeaderComponent,
-    PostListComponent,
-    LoginComponent,
-    SignupComponent,
     ErrorComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    ReactiveFormsModule,
-    FormsModule,
     BrowserAnimationsModule,
-    MatInputModule,
-    MatCardModule,
-    MatButtonModule,
-    MatToolbarModule,
-    MatExpansionModule,
-    MatDialogModule,
-    MatProgressSpinnerModule,
     HttpClientModule,
-    MatPaginatorModule,
-    MatIconModule
+    AngularMetraialModule,
+    PostsModule,
+    AuthModule
   ],
   providers: [
     {provide:  HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true},
