+++ src/app/auth/auth.module.ts
@@ -1,0 +1,23 @@
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+
+import { FormsModule } from "@angular/forms";
+import { AngularMetraialModule } from "../angular-matrial.module";
+
+import { LoginComponent } from './login/login.component';
+import { SignupComponent } from './signup/signup.component';
+
+
+@NgModule({
+  declarations: [
+    LoginComponent,
+    SignupComponent,
+  ],
+  imports: [
+    CommonModule,
+    FormsModule,
+    AngularMetraialModule
+  ]
+})
+
+export class AuthModule {}
